requirements.txt
pydub
streamlit
streamlit-audiorecorder
qdrant-client
openai
python-dotenv